calculate star bg adu by fitting bg change fn over x
